requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
yfinance>=0.2.43
anthropic==0.42.0
python-dotenv==1.0.1
aiosqlite==0.20.0
httpx==0.27.2
apscheduler==3.10.4
sse-starlette==2.1.3
beautifulsoup4==4.12.3
tweepy>=4.14.0

